.htaccess: redirect ipv4 and ipv6 subdomains
Until now I have allowed accessing the site via ipv4.kevinlocke.name and
ipv6.kevinlocke.name.  Although this can be helpful for troubleshooting,
it creates crawling and search indexing issues: crawlers can not be
certain the content is an exact duplicate and which is the canonical
URL.  Although I could [noindex] it, or block crawling with
[robots.txt] (after removing it from the search index), the marginal
troubleshooting value doesn't justify the hassle and potential issues
with other crawlers.  Instead, redirect to the top-level domain like the
other subdomains (except testing.kevinlocke.name, which is
password-protected) excluding /client-ip, which is useful for testing.

[noindex]: https://developers.google.com/search/docs/crawling-indexing/block-indexing
[robots.txt]: https://developers.google.com/search/docs/crawling-indexing/robots/robots_txt

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>